Refactor: Modularize core analysis, plotting, and data functions
- Created `src/stock_analysis` directory to house new modules.
- Moved `analyze_fixed_time_lag` to `src/stock_analysis/core.py`.
- Moved plotting functions to `src/stock_analysis/plotting.py`.
- Moved `download_stock_data` to `src/stock_analysis/data.py`.
- Updated `dynamic.py` to import and use the new modules.
